LayoutTests/http/tests/security/xssAuditor/base-href-expected.txt

CONSOLE MESSAGE: line 1: Refused to load from document base URL. URL found within request.

ALERT: This is a safe script.
